Refactor(plugins): Move jinja filter code for arista.avd.natural_sort to PyAVD
