+++ python-avd/pyavd/templater.py
@@ -5,6 +5,7 @@ from jinja2 import ChoiceLoader, Environment, FileSystemLoader, ModuleLoader, St
 
 from .constants import JINJA2_EXTENSIONS, JINJA2_PRECOMPILED_TEMPLATE_PATH
 from .j2filters.default import default
+from .j2filters.natural_sort import natural_sort
 
 
 class Undefined(StrictUndefined):
@@ -57,7 +58,6 @@ class Templar:
         from .vendor.j2.filter.encrypt import encrypt
         from .vendor.j2.filter.hide_passwords import hide_passwords
         from .vendor.j2.filter.list_compress import list_compress
-        from .vendor.j2.filter.natural_sort import natural_sort
         from .vendor.j2.filter.range_expand import range_expand
         from .vendor.j2.test.contains import contains
         from .vendor.j2.test.defined import defined
